Move previous tests into creation/methods folder and create creation/classes folder with Tensor.ts test
