+++ apitests/src/creation/classes/Tensor.ts
@@ -1,9 +1,9 @@
 import * as chai from "chai";
 const expect = chai.expect;
-import { tensorChaiPlugin } from "../plugins/tensor-chai";
+import { tensorChaiPlugin } from "../../plugins/tensor-chai";
 chai.use(tensorChaiPlugin);
 import type tfTypes from "@tensorflow/tfjs-core";
-import * as loader from "../load-tf";
+import * as loader from "../../load-tf";
 
 let tf: loader.TFModule;
 
@@ -26,24 +26,6 @@ export function run() {
     const expected = [
       [4, 6],
       [8, 10],
-    ];
-    expect(buffer.toTensor().arraySync()).to.eql(expected);
-  });
-  it("  -- dtype", () => {
-    const buffer: tfTypes.TensorBuffer<tfTypes.Rank.R2, "int32"> = tf.buffer(
-      [2, 2],
-      "int32"
-    );
-    expect(buffer).to.haveDtype("int32");
-  });
-  it("  -- values", () => {
-    const buffer: tfTypes.TensorBuffer<
-      tfTypes.Rank.R2,
-      keyof tfTypes.DataTypeMap
-    > = tf.buffer([2, 2], undefined, Int32Array.from([1, 2, 3, 4]));
-    const expected = [
-      [1, 2],
-      [3, 4],
     ];
     expect(buffer.toTensor().arraySync()).to.eql(expected);
   });
