+++ apitests/src/creation/methods/linspace.ts
@@ -1,10 +1,9 @@
 import * as chai from "chai";
 const expect = chai.expect;
-import { tensorChaiPlugin } from "../plugins/tensor-chai";
+import { tensorChaiPlugin } from "../../plugins/tensor-chai";
 chai.use(tensorChaiPlugin);
 import type tfTypes from "@tensorflow/tfjs-core";
-import * as loader from "../load-tf";
-
+import * as loader from "../../load-tf";
 let tf: loader.TFModule;
 
 export function run() {
