+++ apitests/src/creation/methods/real.ts
@@ -1,9 +1,9 @@
 import * as chai from "chai";
 const expect = chai.expect;
-import { tensorChaiPlugin } from "../plugins/tensor-chai";
+import { tensorChaiPlugin } from "../../plugins/tensor-chai";
 chai.use(tensorChaiPlugin);
 import type tfTypes from "@tensorflow/tfjs-core";
-import * as loader from "../load-tf";
+import * as loader from "../../load-tf";
 
 let tf: loader.TFModule;
