+++ apitests/src/creation/methods/zeros.ts
@@ -1,9 +1,9 @@
 import * as chai from "chai";
 const expect = chai.expect;
-import { tensorChaiPlugin } from "../plugins/tensor-chai";
+import { tensorChaiPlugin } from "../../plugins/tensor-chai";
 chai.use(tensorChaiPlugin);
 import type tfTypes from "@tensorflow/tfjs-core";
-import * as loader from "../load-tf";
+import * as loader from "../../load-tf";
 
 let tf: loader.TFModule;
 
@@ -15,13 +15,10 @@ export function run() {
       done();
     });
   });
-  it("  -- clones a tensor", () => {
-    const arr = [
-      [1, 2, 3],
-      [4, 5, 6],
-    ];
-    const t: tfTypes.Tensor = tf.tensor(arr);
-    const clone: tfTypes.Tensor = t.clone();
-    expect(clone).to.haveShape([2, 3]);
+
+  // TESTS
+  it("  -- default", async () => {
+    const t: tfTypes.Tensor = tf.zeros([2, 2]);
+    expect(t).to.be.allZeros;
   });
 }
